+++ src/lib.rs
@@ -29,6 +29,7 @@ mod td;
 mod vcell;
 
 pub mod full_speed;
+pub mod gpt;
 
 /// Eight endpoints, two directions
 const QH_COUNT: usize = 8 * 2;