+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Changelog
 [Unreleased]
 ------------
 
+- Add support for USB general purpose timers (GPT).
 - Fix the endpoint initialization routine, which would incorrectly zero the
   other half's endpoint type.
-- Fix documentation of `full_speed::BusAdapter::new`
+- Fix documentation of `full_speed::BusAdapter::new`.
 
 [0.1.0] 2021-03-11
 ------------------